moved the custom rules loader to package
