+++ packages/functions/package.json
@@ -22,7 +22,8 @@
   "scripts": {
     "build": "tsc",
     "dev": "tsc -w",
-    "test": "jest --coverage"
+    "test": "jest --coverage",
+    "test:watch": "jest --watch"
   },
   "repository": {
     "type": "git",
