+++ packages/jest-presets/jest/node/jest-preset.js
@@ -1,0 +1,9 @@
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testPathIgnorePatterns: ['dist/'],
+  transform: {
+    '^.+\\.ts?$': 'ts-jest',
+  },
+};
