+++ packages/rulesets/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@compliancepal/spectral-rulesets",
+  "version": "1.0.0",
+  "description": "Custom Spectral rulesets",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "dependencies": {
+    "@stoplight/json": "^3.20.1",
+    "@stoplight/path": "^1.3.2",
+    "@stoplight/spectral-core": "^1.15.0",
+    "@stoplight/spectral-formats": "^1.4.0",
+    "@stoplight/spectral-functions": "^1.7.1",
+    "@stoplight/spectral-parsers": "^1.0.2",
+    "@stoplight/spectral-ref-resolver": "^1.0.2",
+    "@stoplight/spectral-ruleset-bundler": "^1.3.2",
+    "@stoplight/spectral-ruleset-migrator": "^1.8.0",
+    "@stoplight/spectral-rulesets": "^1.14.1",
+    "@stoplight/spectral-runtime": "^1.1.2",
+    "@stoplight/yaml": "^4.2.3",
+    "rollup": "^2.79.1"
+  },
+  "devDependencies": {
+    "jest": "^28.1.3",
+    "jest-presets": "*",
+    "ts-jest": "^28.0.8",
+    "tsconfig": "*",
+    "typescript": "^4.8.4"
+  },
+  "scripts": {
+    "build": "tsc",
+    "test": "jest --coverage"
+  },
+  "jest": {
+    "preset": "jest-presets/jest/node"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/CompliancePal/modelcard-action.git"
+  },
+  "author": "Vlad Știrbu",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/CompliancePal/modelcard-action/issues"
+  },
+  "homepage": "https://github.com/CompliancePal/modelcard-action#readme"
+}
