+++ packages/rulesets/src/__fixtures__/rules.yaml
