+++ packages/rulesets/src/index.ts
@@ -1,0 +1,1 @@
+export { default as loader } from './rule-loader';
